py/packages/corpora/requirements.txt
Django==5.1.2
django-ninja==1.3.0
pgvector==0.3.6
psycopg2==2.9.10
